requirements.aws.txt
aws-cdk.core
aws-cdk.aws_iam
aws-cdk.aws_lambda
aws-cdk.aws_s3
aws-cdk.aws_s3_notifications
aws-cdk.aws_sns
aws-cdk.aws_sns_subscriptions
